simplifications to the typing of basefluent
fixes fabric8io/kubernetes-client#5522
