--- core/src/main/java/io/sundr/builder/Fluent.java
@@ -1,27 +1,0 @@
-/*
- * Copyright 2015 The original authors.
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
-package io.sundr.builder;
-
-/**
- * The base interface for Fluent interfaces.
- * Implementers of this interface are usually classes that need to provide chained method calls (e.g. Builders).
- * 
- * @param <F> The type that is expected to be returned by the Fluent methods.
- */
-public interface Fluent<F extends Fluent<F>> {
-
-}
